requirements.txt
pandas~=1.3.5
beautifulsoup4~=4.11.1
numpy~=1.19.5
tensorflow-gpu~=2.6.0
transformers~=4.18.0
tqdm~=4.64.0
requests~=2.27.1
flask~=2.2.1
easydict~=1.9
pymongo~=4.2.0
seqeval~=1.2.2
torch
scikit-learn~=0.22.1
zhconv~=1.4.3
bert4keras~=0.8.1
gensim~=3.7.3=
keras~=2.6.0
keras_self_attention
keras_multi_head
typeguard
tensorflow_addons